Making attempts to resolve the HARD_FAULT crashes
Made some minor adjustments to try to fix the Hard fault crashes. While doing that I noticed that there were a couple of communication issues related to the receive buffer. The rx buffer was design manually (static) only able to add more data if all the data had been read, however, since the balancer algorithm command is about 54 bytes, it periodically gets segmented and returns to the UI as unrecognized commands. This is even worse since a balance request may not be caught.
I'm going to try to switch this static buffer to a dynamic buffer.

The other issue was the Hello response used to verify communication had multiple newlines. Likewise, the "Not a recognized command" was improved to include the command it had checked.