Changed parser command request from buffer to peek instead
This is a part to the solution to the "HardFault" error. The circularbuffer was the other complementary solution. In this case the circularbuffer is used in this case to allow data to be peeked and to allow a FIFO cycle. The problem was an overflow problem whenever a long series of bytes was written to the previous buffer. Before, the "usb_parser_command" would be removed from the previous arraybuffer, and then the rest of the code would need to evaluate the rest of it. But in a case where the "rest of the data" was not yet received due to the array being full, the parser would try extracting data outside the range of the array. The newly implemented circularbuffer allows the index of data to be moved once removed, thereby allowing longer streams to be received. But the main point to this commit is the fact that the "usb_parser_command" is peeked at, and then removed by the parsers only when the required amount of data is available (see "case USB_PARSER_ALGORITHM_COMMAND"